Add quantity

+++ model/prfDetailsModel.js
@@ -11,7 +11,7 @@ const savePRFDetails = async (prfDetailsArray) => {
         table.columns.add("StockCode", sql.VarChar(250), { nullable: false });
         table.columns.add("StockName", sql.VarChar(sql.MAX), { nullable: false });
         table.columns.add("UOM", sql.VarChar(100), { nullable: false });
-        table.columns.add("QTY", sql.Int, { nullable: false });
+        table.columns.add("QTY", sql.Decimal(10,2), { nullable: false });
         table.columns.add("DateNeeded", sql.VarChar(250), { nullable: false });
         table.columns.add("Purpose", sql.VarChar(250), { nullable: false });
         table.columns.add("Description", sql.VarChar(250), { nullable: true });
